rename files and clean more unused functions
