+++ cosmos_predict2/configs/action_conditional/defaults/model.py
@@ -17,7 +17,7 @@ from hydra.core.config_store import ConfigStore
 from cosmos_predict2.configs.action_conditional.config_action_conditional import (
     ACTION_CONDITIONAL_PREDICT2_VIDEO2WORLD_PIPELINE_2B,
 )
-from cosmos_predict2.models.video2world_model_action import (
+from cosmos_predict2.models.action_video2world_model import (
     ActionConditionalPredict2Video2WorldModel,
 )
 from cosmos_predict2.models.video2world_model import Predict2ModelManagerConfig, Predict2Video2WorldModelConfig
